+++ tests/tdl_test.py
@@ -1008,3 +1008,20 @@ def test_issue_357():
     assert isinstance(c, ConsList)
     assert len(c.values()) == 2
     assert tdl.format(t) == 'id := < a . < b . c > >.'
+
+
+def test_promote_conjunction_issue_395():
+    t = tdlparse('a := b & [ ABC x, ABC [ DEF y ] ].')
+    assert tdl.format(t) == 'a := b &\n  [ ABC x & [ DEF y ] ].'
+    t = tdlparse('a := b & [ ABC x, ABC.DEF y ].')
+    assert tdl.format(t) == 'a := b &\n  [ ABC x & [ DEF y ] ].'
+    t = tdlparse('a := b & [ ABC.DEF y, ABC x ].')
+    assert tdl.format(t) == 'a := b &\n  [ ABC [ DEF y ] & x ].'
+    t = tdlparse('a := b & [ ABC [ DEF x ], ABC [ GHI y ] ].')
+    assert tdl.format(t) == 'a := b &\n  [ ABC [ DEF x ] & [ GHI y ] ].'
+    t = tdlparse('a := b & [ ABC [ DEF x ], ABC.DEF y ].')
+    assert tdl.format(t) == 'a := b &\n  [ ABC [ DEF x ] & [ DEF y ] ].'
+    t = tdlparse('a := b & [ ABC [ DEF x ], ABC.GHI y ].')
+    assert tdl.format(t) == 'a := b &\n  [ ABC [ DEF x ] & [ GHI y ] ].'
+    t = tdlparse('a := b & [ ABC x & [ DEF y ], ABC z ].')
+    assert tdl.format(t) == 'a := b &\n  [ ABC x & [ DEF y ] & z ].'